Allow custom AuxiliaryLightManager implementations apart from the chunk-bound one
